+++ NEWS.md
@@ -1,3 +1,15 @@
+## alfr 1.2.0
+
+* Added folder and document delete
+* Added session invalidate
+* Added session is valid check
+* Use pkgdown to generate package site documentation
+* Improved examples
+* Unit test refactor
+   - tests atomic
+   - mocked API easy to refresh
+   - coverage > 90%
+
 ## alfr 1.1.0
 
 * Improved documentation